refactor: move TextMessageItemBody to ClampedMessageItemBody
