+++ src/rogu/ui/MessageContent/index.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement, useRef, useContext } from 'react';
 import { GroupChannel, AdminMessage, UserMessage, FileMessage } from 'sendbird';
 import Label, { LabelTypography, LabelColors } from '../Label';
 import MessageStatus from '../MessageStatus';
-import TextMessageItemBody from '../TextMessageItemBody';
+import ClampedMessageItemBody from '../ClampedMessageItemBody';
 import OGMessageItemBody from '../OGMessageItemBody';
 import FileMessageItemBody from '../FileMessageItemBody';
 import ThumbnailMessageItemBody from '../ThumbnailMessageItemBody';
@@ -83,8 +83,8 @@ export default function MessageContent({
   resendMessage,
   disabled = false,
 }: // showRemove,
-  // toggleReaction,
-  Props): ReactElement {
+// toggleReaction,
+Props): ReactElement {
   const { stringSet } = useContext(LocalizationContext);
   const messageTypes = getUIKitMessageTypes();
   const avatarRef = useRef(null);
@@ -98,8 +98,6 @@ export default function MessageContent({
     message as CoreMessageType
   );
 
-
-
   const isByMeClassName = isByMe
     ? 'rogu-message-content--outgoing'
     : 'rogu-message-content--incoming';
@@ -111,7 +109,6 @@ export default function MessageContent({
   if (message?.isAdminMessage?.() || message?.messageType === 'admin') {
     return <ClientAdminMessage message={message} />;
   }
-
 
   const onScrollToMessage = () => {
     //TODO: integrate onScrollToMessage
@@ -168,7 +165,7 @@ export default function MessageContent({
                     {stringSet.LABEL__OPERATOR}
                   </Label>
                 )}
-                {!channel.isFrozen &&
+                {!channel.isFrozen && (
                   <MessageItemMenu
                     className="rogu-message-content__menu"
                     channel={channel}
@@ -180,7 +177,7 @@ export default function MessageContent({
                     resendMessage={resendMessage}
                     showFileViewer={showFileViewer}
                   />
-                }
+                )}
               </>
             )}
           </div>
@@ -188,11 +185,9 @@ export default function MessageContent({
             <div className="rogu-message-content__bubble__body__inner">
               {/* Message content */}
               {isTextMessage(message as UserMessage) && (
-                <TextMessageItemBody
+                <ClampedMessageItemBody
                   isByMe={isByMe}
-                  message={(message as UserMessage)?.message}
-                  isRepliedMessage={isRepliedMessage(message)}
-                  onScrollToMessage={onScrollToMessage}
+                  content={(message as UserMessage)?.message}
                 />
               )}
               {isOGMessage(message as UserMessage) && (
@@ -215,11 +210,11 @@ export default function MessageContent({
               )}
               {getUIKitMessageType(message as FileMessage) ===
                 messageTypes.FILE && (
-                  <FileMessageItemBody
-                    message={message as FileMessage}
-                    isByMe={isByMe}
-                  />
-                )}
+                <FileMessageItemBody
+                  message={message as FileMessage}
+                  isByMe={isByMe}
+                />
+              )}
               {isThumbnailMessage(message as FileMessage) && (
                 <>
                   <ThumbnailMessageItemBody
@@ -232,18 +227,18 @@ export default function MessageContent({
                     }
                   />
                   {(message as FileMessage).name && (
-                    <TextMessageItemBody
+                    <ClampedMessageItemBody
                       isByMe={isByMe}
                       mode="thumbnailCaption"
-                      message={(message as FileMessage).name}
+                      content={(message as FileMessage).name}
                     />
                   )}
                 </>
               )}
               {getUIKitMessageType(message as FileMessage) ===
                 messageTypes.UNKNOWN && (
-                  <UnknownMessageItemBody message={message} isByMe={isByMe} />
-                )}
+                <UnknownMessageItemBody message={message} isByMe={isByMe} />
+              )}
             </div>
             {((!isByMe && chainTop) || isByMe) && (
               <MessageItemMenu
