--- src/rogu/ui/TextMessageItemBody/index.scss
@@ -1,98 +1,0 @@
-@import '../../../styles/variables';
-
-.rogu-text-message-item-body {
-  position: relative;
-  display: inline-block;
-  box-sizing: content-box;
-
-  .rogu-text-message-item-body__message {
-    white-space: pre-wrap;
-    word-break: break-all;
-  }
-
-  // .rogu-text-message-item-body__reply-container {
-  //   padding: 8px;
-  //   border-radius: 4px;
-  //   margin-bottom: 8px;
-  //   border-left-width: 2px;
-  //   border-left-style: solid;
-  //   cursor: pointer;
-
-  //   @include themed() {
-  //     background-color: t(primary-2);
-  //     border-left-color: t(primary-5);
-  //   }
-  // }
-
-  .rogu-text-message-item-body__reply-message {
-    white-space: pre-wrap;
-    word-break: break-all;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-  }
-
-  .rogu-text-message-item-body__read-more {
-    font-weight: bolder;
-  }
-
-  .rogu-text-message-item-body__inner {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: -webkit-box;
-    -webkit-line-clamp: 10;
-    -webkit-box-orient: vertical;
-    align-self: center;
-  }
-
-  &.rogu-text-message-item-body--incoming {
-    .rogu-text-message-item-body__read-more {
-      @include themed() {
-        color: t(secondary-3);
-      }
-    }
-
-    .rogu-replied-text-message-item-body__reply-container {
-      @include themed() {
-        background-color: t(bg-1);
-      }
-    }
-  }
-
-  &.rogu-text-message-item-body--expanded {
-    .rogu-text-message-item-body__inner {
-      display: block;
-      overflow: visible;
-    }
-  }
-
-  &.rogu-text-message-item-body--viewer-mode {
-    position: absolute;
-    bottom: 5%;
-    max-height: 40%;
-    min-width: 85%;
-    max-width: 95%;
-    background: #f1f4f9;
-    backdrop-filter: blur(16px);
-    padding: 24px;
-    overflow: auto;
-    transition: all 0.2s;
-    display: flex;
-    flex-direction: column;
-
-    @include sm {
-      max-width: 70%;
-    }
-
-    &.rogu-text-message-item-body--viewer-mode__hidden {
-      max-height: 0;
-      padding: 0;
-    }
-  }
-
-  &.rogu-text-message-item-body--preview-mode {
-    display: block;
-  }
-}
